% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenDummy.R
\name{gen_dummy}
\alias{gen.dummy}
\alias{gen_dummy}
\title{Generate dummy variables from ONE variable}
\usage{
gen.dummy(vec, levels = NULL, prefix = NULL, head = NULL, bool = FALSE)
}
\arguments{
\item{vec}{variable vector}

\item{levels}{variable values, each generates a dummy}

\item{prefix}{prefix for dummy variables}

\item{head}{dummy variable names}

\item{bool}{return bool dummy variables if TRUE}
}
\value{
dummy variable matrix
}
\description{
Generate dummy variables from ONE variable
}

