% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximunLikelihoodEstimation.R
\name{maximun_likelihood_estimation}
\alias{maximun_likelihood_estimation}
\alias{mle.model}
\title{Maximun Likelihood Estimation Model}
\usage{
mle.model(lnlike, args, params0, gr = NULL, method = "BFGS")
}
\arguments{
\item{lnlike}{log-likelihood function}

\item{args}{fixed arguments for lnlike and gr, usually are data to fit. 
\tiem{x} explanatory variable
\item{y} explained variable}

\item{params0}{initial parameters to optimize}

\item{gr}{gradient function}

\item{method}{optimization method}
}
\value{
a list with 4 elements
\item{convergence} 
\item{observations} 
\item{lnlikelihood} 
\item{table} regression result table
}
\description{
Base for MLE models
}
\author{
syuoni
}

