% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE-BinaryModels.R
\name{mle_binary_models}
\alias{mle.logit.estimate}
\alias{mle.probit.estimate}
\alias{mle_binary_models}
\title{Binary Models Estimation with MLE}
\usage{
mle.probit.estimate(y, X)

mle.logit.estimate(y, X)
}
\arguments{
\item{y}{explained variable, binary}

\item{X}{explanatory variable in matrix}
}
\value{
a list with 4 elements, mle results
\item{convergence} 
\item{observations} 
\item{lnlikelihood} 
\item{table} regression result table
}
\author{
syuoni
}

