% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE-DurationModels.R
\name{mle_duration_models}
\alias{mle.exp.estimate}
\alias{mle.gompertz.estimate}
\alias{mle.weibull.estimate}
\alias{mle_duration_models}
\title{Duration Models Estimation with MLE}
\usage{
mle.exp.estimate(t, X, d)

mle.weibull.estimate(t, X, d)

mle.gompertz.estimate(t, X, d)
}
\arguments{
\item{t}{explained variable, vector, duration time}

\item{X}{explanatory variable in matrix}

\item{d}{explanatory variable, vector, whether the sample is failed (1=failure)}
}
\value{
a list with 4 elements, mle result
  \item{convergence}{1 if converge successfully, 0 if fail}
  \item{observations}{number of samples}
  \item{lnlikelihood}{log(likelihood)}
  \item{table}{regression result table}
}
\description{
Duration Models Estimation with MLE
}
\author{
syuoni
}

