% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE-DurationModelsWithSelection.R
\name{mle_duration_models_with_selection}
\alias{mle.exp.selection.estimate}
\alias{mle.weibull.selection.estimate}
\alias{mle_duration_models_with_selection}
\title{Duration Models With Selection Estimation with MLE}
\usage{
mle.exp.selection.estimate(t, X, z, W, d)

mle.weibull.selection.estimate(t, X, z, W, d)
}
\arguments{
\item{t}{explained variable, vector, duration time}

\item{X}{explanatory variable in matrix}

\item{z}{selection explained variable, binary}

\item{W}{selection explanatory variable in matrix}

\item{d}{explanatory variable, vector, whether the sample is failed (1=failure)}
}
\value{
a list with 5 elements, mle result
  \item{method}{'mle'}
  \item{convergence}{1 if converge successfully, 0 if fail}
  \item{observations}{number of samples}
  \item{lnlikelihood}{log(likelihood)}
  \item{table}{regression result table}
}
\description{
Duration Models With Selection Estimation with MLE
}
\references{
Boehmke F J, Morey D S, Shannon M. Selection Bias and Continuous Time Duration Models: 
  Consequences and a Proposed Solution[J]. American Journal of Political Science, 2006, 50(1): 192-207.
}

