% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE-LinearModels.R
\name{mle_linear_models}
\alias{mle.linear.estimate}
\alias{mle_linear_models}
\title{Linear Models Estimation with MLE}
\usage{
mle.linear.estimate(y, X)
}
\arguments{
\item{y}{explained variable}

\item{X}{explanatory variable in matrix}
}
\value{
a list with 4 elements, mle results
\item{convergence} 
\item{observations} 
\item{lnlikelihood} 
\item{table} regression result table
}
\author{
syuoni
}

