% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE-ModelResultExport.R
\name{mle_model_res_export}
\alias{mle.res.export}
\alias{mle.res.table.export}
\alias{mle_model_res_export}
\title{model results export}
\usage{
mle.res.export(model.res, ...)

mle.res.table.export(model.res.list, all.params = NULL, ...)
}
\arguments{
\item{model.res}{ONE mle regression model result}

\item{model.res.list}{mle regression models results}
}
\value{
a vector

a data frame
}
\description{
export ONE mle model as a vector

export MULTIPLE mle models as a table (data frame)
}
\author{
syuoni
}

