% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelResultExport.R
\name{model_res_export}
\alias{model.res.export}
\alias{model_res_export}
\title{Export ONE regression model result as a vector.}
\usage{
model.res.export(model.res, all.params = NULL,
  parentheses.type = "t.statistic", digits = 3)
}
\arguments{
\item{model.res}{ONE regression model result}

\item{all.params}{all potential parameters}

\item{parentheses.type}{'t-statistic' or 'standard error' in parentheses}

\item{digits}{digits}
}
\value{
a vector of regression model result
}
\description{
Export ONE regression model result as a vector.
}

