% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelResultTableExport.R
\name{model_res_table_export}
\alias{model.res.table.export}
\alias{model_res_table_export}
\title{export MULTIPLE models as a table (data frame)}
\usage{
model.res.table.export(model.res.list, all.params = NULL, ...)
}
\arguments{
\item{model.res.list}{a list of several regression model results}

\item{all.params}{all potential parameters}

\item{...}{parameters passed to function model_res_export}
}
\value{
a data frame
}
\description{
export MULTIPLE models as a table (data frame)
}

