% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericGradient.R
\name{numeric_gr}
\alias{numeric.gr}
\alias{numeric_gr}
\title{Generage a numeric gradient function}
\usage{
numeric.gr(func)
}
\arguments{
\item{func}{function to calculate numeric gradient function}
}
\value{
gradient function
}
\description{
Generage a numeric gradient function
}

