% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericGradient.R
\name{numeric_gradient}
\alias{numeric.gr}
\alias{numeric_gradient}
\title{Numeric Gradient Function}
\usage{
numeric.gr(func)
}
\arguments{
\item{func}{function to calculate numeric gradient function}
}
\value{
gradient function
}
\description{
Numeric Gradient Function
}
\author{
syuoni
}

