% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS-WithinEstimate.R
\name{ols_within_estimate}
\alias{ols.within.estimate}
\alias{ols_within_estimate}
\title{Within Estimate}
\usage{
ols.within.estimate(y, X, group.indic, robust = FALSE)
}
\arguments{
\item{y}{explained variable}

\item{X}{explanatory variable in matrix}

\item{group.indic}{group indicator variable}

\item{robust}{whether use robust convariance}
}
\value{
a list with 8 elements
  \item{method}{'ols'}
  \item{convariance}{1 if robust, 0 if not}
  \item{observations}{number of samples}
  \item{R.square}{R^2}
  \item{adj.R.square}{adjusted R^2}
  \item{F.statistic}{F-statistic}
  \item{Prob.F}{Prob(F)}
  \item{table}{regression result table}
}
\description{
Within Estimate
}

