% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticR.R
\docType{package}
\name{statistic_r}
\alias{statistic_r}
\alias{statistic_r-package}
\title{An implementation of statistic models with R.}
\description{
An implementation of statistic models with R. 
  This package contains basically Econometrics models.
}

