% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStep-Heckman.R
\name{twostep_heckman}
\alias{twostep.heckman.estimate}
\alias{twostep_heckman}
\title{Heckman Models Estimation with Two Step, i.e. Heckit}
\usage{
twostep.heckman.estimate(y, X, z, W)
}
\arguments{
\item{y}{explained variable}

\item{X}{explanatory variable in matrix}

\item{z}{selection explained variable, binary}

\item{W}{selection explanatory variable in matrix}
}
\value{
a list with 5 elements
  \item{method}{'two-step'}
  \item{selection}{selection model result of probit}
  \item{regression}{regression model result of ols}
  \item{sigma}{sigma}
  \item{rho}{rho}
}
\description{
Heckman Models Estimation with Two Step, i.e. Heckit
}

