% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean4Regression.R
\name{clean4regression}
\alias{clean4regression}
\title{Clean data frame for fegression}
\usage{
clean4regression(args, assign.labels = NULL, exempt.cond = FALSE,
  add.const = TRUE)
}
\arguments{
\item{args}{arguments for regression model, before cleaning}

\item{assign.labels}{NA samples in args with specific labels would be dropped}

\item{exempt.cond}{vector, exempt condition, sample would not be dropped if it is TRUE}

\item{add.const}{whether add const variable for matrix}
}
\value{
a list, arguments for regression model, after cleaning
}
\description{
Clean data frame for fegression, including dropping NA observations 
  and adding constant variable, make it prepared for regression models
}

