% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenDummy.R
\name{gen_dummy}
\alias{gen.dummy}
\alias{gen_dummy}
\title{Generate Dummy}
\usage{
gen.dummy(vec, levels = NULL, head = NULL)
}
\arguments{
\item{vec}{variable vector}

\item{levels}{variable values, each generates a dummy}

\item{head}{dummy variable names}
}
\value{
dummy variable matrix
}
\author{
syuoni
}

