% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximunLikelihoodEstimation.R
\name{maximun_likelihood_estimation}
\alias{maximun_likelihood_estimation}
\alias{mle.model}
\title{Maximun Likelihood Estimation(MLE) Model, the base for MLE models}
\usage{
mle.model(lnlike, args, params0, gr = NULL, method = "BFGS")
}
\arguments{
\item{lnlike}{log-likelihood function}

\item{args}{A list. Fixed arguments for lnlike and gr, usually are data to fit. 
Usually include y as explained variable, X as explanatory variable}

\item{params0}{initial parameters to optimize}

\item{gr}{gradient function}

\item{method}{optimization method}
}
\value{
a list with 5 elements
  \item{method}{'mle'}
  \item{convergence}{1 if converge successfully, 0 if fail}
  \item{observations}{number of samples}
  \item{lnlikelihood}{log(likelihood)}
  \item{table}{regression result table}
}
\description{
Maximun Likelihood Estimation(MLE) Model, the base for MLE models
}

