% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE-Heckman.R
\name{mle_heckman}
\alias{mle.heckman.estimate}
\alias{mle_heckman}
\title{Heckman Models Estimation with MLE, Heckit}
\usage{
mle.heckman.estimate(y, X, z, W)
}
\arguments{
\item{y}{explained variable}

\item{X}{explanatory variable in matrix}

\item{z}{selection explained variable, binary}

\item{W}{selection explanatory variable in matrix}
}
\value{
a list with 4 elements, mle results
\item{convergence} 
\item{observations} 
\item{lnlikelihood} 
\item{table} regression result table
}
\author{
syuoni
}

