% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE-Heckman.R
\name{mle_heckman}
\alias{mle.heckman.estimate}
\alias{mle_heckman}
\title{Heckman Models Estimation with MLE, Heckit}
\usage{
mle.heckman.estimate(y, X, z, W)
}
\arguments{
\item{y}{explained variable}

\item{X}{explanatory variable in matrix}

\item{z}{selection explained variable, binary}

\item{W}{selection explanatory variable in matrix}
}
\value{
a list with 5 elements, mle result
  \item{method}{'mle'}
  \item{convergence}{1 if converge successfully, 0 if fail}
  \item{observations}{number of samples}
  \item{lnlikelihood}{log(likelihood)}
  \item{table}{regression result table}
}
\description{
Heckman Models Estimation with MLE, Heckit
}

