% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS-ModelResultExport.R
\name{ols_model_res_export}
\alias{ols.res.export}
\alias{ols_model_res_export}
\title{model results export for OLS}
\usage{
ols.res.export(model.res, ...)
}
\arguments{
\item{model.res}{ONE ols regression model result}

\item{...}{parameters passed}
}
\value{
a vector
}
\description{
export ONE ols model as a vector
}
\author{
syuoni
}

