% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinaryLeastSquare.R
\name{ordinary_least_square}
\alias{ols.estimate}
\alias{ordinary_least_square}
\title{Ordinary Least Square(OLS) Model}
\usage{
ols.estimate(y, X, robust = FALSE, add.const = TRUE)
}
\arguments{
\item{y}{explained variable}

\item{X}{explanatory variable in matrix}

\item{robust}{whether use robust convariance}

\item{add.const}{whether add constant variable}
}
\value{
a list with 8 elements
  \item{method}{'ols'}
  \item{robust}{use robust covariance if TRUE}
  \item{observations}{number of samples}
  \item{R.square}{R^2}
  \item{adj.R.square}{adjusted R^2}
  \item{F.statistic}{F-statistic}
  \item{Prob.F}{Prob(F)}
  \item{table}{regression result table}
}
\description{
Ordinary Least Square(OLS) Model
}

