% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinaryLeastSquare.R
\name{ordinary_least_square}
\alias{ols.estimate}
\alias{ordinary_least_square}
\title{Ordinary Least Square}
\usage{
ols.estimate(y, X, robust = FALSE)
}
\arguments{
\item{y}{explained variable}

\item{X}{explanatory variable in matrix}
}
\value{
a list with 7 elements, mle results
\item{convariance} whether robust
\item{observations} 
\item{R.square} R^2
\item{adj.R.square} adjusted R^2
\item{F.statistic} F-statistic
\item{Prob.F} Prob(F)
\item{table} regression result table
}
\author{
syuoni
}

